requirements.txt
matplotlib==3.3.4
pandas==1.2.1
dnspython==2.1.0
Requests==2.25.1
urllib3==1.26.3